added content crud and category and blog model

+++ administrator/views/group.py
@@ -1,12 +1,7 @@
-import ast
-
 from django.http import HttpResponse
 from django.shortcuts import redirect, render, get_object_or_404
-
-from Users.models import CustomUser
-from administrator.forms import UserForm, UserUpdateForm, GroupForm
+from administrator.forms import GroupForm
 from django.db.models import Q
-
 from django.core.paginator import Paginator
 from administrator.services.users import UserService
 from administrator.views.base import BaseAdminView
@@ -15,12 +10,14 @@ import json
 from django.contrib.auth.models import Group
 
 
-class GroupsView(BaseAdminView):
+class GroupView(BaseAdminView):
     def dispatch(self, request, *args, **kwargs):
+        request.session['menu'] = 'Group'
         return super().dispatch(request, *args, **kwargs, session_menu='Group', session_submenu='',
                                 permission_required=[])
 
-    def get(self, request):
+    @classmethod
+    def get_list(cls, request):
         keyword = request.GET.get('keyword')
         filter = {'keyword': keyword if keyword is not None else ""}
         if keyword is None:
@@ -34,41 +31,61 @@ class GroupsView(BaseAdminView):
                       {'paginator': paginator, 'page_number': page_number, 'page_obj': page_obj,
                        'filter': filter})
 
-
-class GroupView(BaseAdminView):
-    def dispatch(self, request, *args, **kwargs):
-        request.session['menu'] = 'Group'
-        return super().dispatch(request, *args, **kwargs, session_menu='User', session_submenu='',
-                                permission_required=[])
-
     @classmethod
-    def add_group(cls, request):
-        group = Group()
-        group_form = GroupForm(instance=group)
-        print(group_form)
-        group.id = 0
-        user_service = UserService()
-        post_data = user_service.getPostData(vars(group), None)
-        return render(request, 'admin/groups/add_group.html',
-                      {'postData': post_data, 'group_permissions': None})
+    def add(cls, request):
+        if request.method == 'POST':
+            _post_data = request.POST
+            post_data = _post_data.dict()
+            post_data.pop('csrfmiddlewaretoken')
+            user_service = UserService()
+            group = Group()
+            post_form = GroupForm(post_data, instance=group)
+            if post_form.is_valid():
+                status = post_form.save()
+                if status:
+                    messages.success(request, 'Success')
+                    return redirect('adminGroupDetail', pk=group.pk)
+                else:
+                    messages.error(request, 'Error occurred while saving group')
+                    post_data['id'] = 0
+                    post_data = user_service.getPostData(post_data, None)
+                    return render(request, 'admin/groups/add_group.html',
+                                  {'postData': post_data, 'permissions': None})
 
-    @classmethod
-    def delete_group(cls, request):
-        _post_data = request.POST
-        pk = _post_data.get('id')
-        group = Group.objects.get(pk=pk)
-        status = group.delete()
-        if status:
-            messages.success(request, 'Group Deleted Successfully')
-            return HttpResponse(json.dumps({'success': True}), content_type="application/json")
+            else:
+                messages.error(request, 'Form validation Error. Please correct the below mentioned errors')
+                print(post_form.errors)
+                errors = json.loads(post_form.errors.as_json())  # errors to json and then to dict
+                post_data = user_service.getPostData(post_data, errors)
+                return render(request, 'admin/groups/add_group.html',
+                              {'postData': post_data, 'permissions': None})
+
         else:
-            messages.error(request, 'There was a problem in deleting the group')
-            return HttpResponse(json.dumps({'success': False}), content_type="application/json")
+            group = Group()
+            group.id = 0
+            user_service = UserService()
+            post_data = user_service.getPostData(vars(group), None)
+            return render(request, 'admin/groups/add_group.html',
+                          {'postData': post_data, 'group_permissions': None})
+
+    @classmethod
+    def delete(cls, request, pk):
+        if request.method == 'POST':
+            _post_data = request.POST
+            group = get_object_or_404(Group, pk=pk)
+            status = group.delete()
+            if status:
+                messages.success(request, 'Group Deleted Successfully')
+                return HttpResponse(json.dumps({'success': True}), content_type="application/json")
+            else:
+                messages.error(request, 'There was a problem in deleting the group')
+                return HttpResponse(json.dumps({'success': False}), content_type="application/json")
+        else:
+            return redirect('adminGroupDetail', pk=pk)
 
     def get(self, request, pk):
-
         if not pk:
-            return redirect('adminAddGroup')
+            return redirect('adminGroupAdd')
         user_service = UserService()
         group = get_object_or_404(Group, pk=pk)
         group_form = GroupForm(instance=group)
@@ -102,13 +119,13 @@ class GroupView(BaseAdminView):
 
             if status:
                 messages.success(request, 'Success')
-                return redirect('adminGroupDetail', pk=pk if pk > 0 else group.pk)
+                return redirect('adminGroupDetail', pk=pk)
             else:
                 messages.error(request, 'Error occurred while saving group')
                 post_data['id'] = pk
                 post_data = user_service.getPostData(post_data, None)
                 return render(request, 'admin/groups/add_group.html',
-                              {'postData': post_data, 'permissions': permissions if pk > 0 else None})
+                              {'postData': post_data, 'permissions': permissions})
 
         else:
             messages.error(request, 'Form validation Error. Please correct the below mentioned errors')
@@ -116,4 +133,4 @@ class GroupView(BaseAdminView):
             errors = json.loads(post_form.errors.as_json())  # errors to json and then to dict
             post_data = user_service.getPostData(post_data, errors)
             return render(request, 'admin/groups/add_group.html',
-                          {'postData': post_data, 'permissions': permissions if pk > 0 else None})
+                          {'postData': post_data, 'permissions': permissions})
